[BEAM-9891] Added ZetaSQL planner support and uploaded 100G data
