+++ sdks/java/testing/tpcds/src/main/java/org/apache/beam/sdk/tpcds/TpcdsSchemas.java
@@ -669,5 +669,4 @@ public class TpcdsSchemas {
                     .addNullableField("web_gmt_offset", Schema.FieldType.DOUBLE)
                     .addNullableField("web_tax_percentage", Schema.FieldType.DOUBLE)
                     .build();
-
 }
